salt: Add new beacon module to check kubeconfig expiry
This beacon takes a list of kubeconfig as input
and checks whether they need to be renewed or not
triggering an event on Salt bus when needed.

To configure this beacon, a section must be
added, either in minion configuration or through
the pillar, as follows:

beacons:
  metalk8s_kubeconfig_info:
    - files:
        - /etc/kubernetes/calico.conf
        - /etc/kubernetes/admin.conf:
            notify_days: 30
    - interval: 86400
    - notify_days: 15

Default notify_days, if not provided, is 45.
It can be overridden for a specific kubeconfig as
shown above.

Refs: #1887