ui/volumes: Fix the bugs cause too many API calls
For non-specified instance URL, such as `/volumes`,
we should automatically choose the first one in the list.

However, since we are not aware currectVolumeName in VolumePage
component, we fall into  update the URL which triger the
VolumePage re-renders. Cause the infinite API calls.

So I move this logic to VolumePageContent which solve the issue.

Refs: #2931